Add compatibility with Lootr for obsidian pillars.
This was a request on my tracker. Unfortunately, it's not possible for
me to hook into the tile entity's `setLootTable` method as the tile has
no context around it allowing me to replace it with a Lootr chest.

Using the LockableLootTileEntity's `setLootTable` static method, on the
other hand, provides all of the ncessary information, and is otherwise
identical to the previous `setLootTable` statement.