Fix feeding trough chunk memory leak, closes #3560
This feature worked by going through every ticking entity in the world
and checking if it is a trough. This used to work decently in the past,
but in 1.18 it had to get the TE from the world position.

This is getting replaced with the vanilla point of interest system.